refactor(swc)!: introduce per-element swc-<tag>.js side-effect entry points
Aligns 2nd-gen import paths with 1st-gen's per-element side-effect convention.
Each migrated component now exposes a dedicated `swc-<tag>.ts` source file that
registers a single custom element, and `index.ts` only re-exports the class.

The canonical side-effect import becomes:

    import '@adobe/spectrum-wc/components/badge/swc-badge.js';

The short-form path (`@adobe/spectrum-wc/badge`) still resolves to the class
for type imports but no longer registers the element.

Changes:
- Add `swc-<tag>.ts` for 12 elements across asset, avatar, badge, button,
  divider, icon, illustrated-message, progress-circle, status-light, and
  tabs (3 elements).
- Refactor each `index.ts` to drop `defineElement` and `HTMLElementTagNameMap`
  (both now live in the matching `swc-<tag>.ts`).
- Extend the Vite library entry glob and switch the package alias to array
  form so the long-form `@adobe/spectrum-wc/components/<comp>/...` path
  resolves cleanly alongside the short-form alias.
- Add `./components/*.js` to package.json `exports` so the published package
  exposes the new side-effect entries.
- Migrate in-repo callers (tests, stories, conversational-ai patterns) to
  the long-form path. Tabs callers now import three side-effect entries
  (one per element) instead of one.
- Update all component migration guides and the `Learn about SWC / Get started`
  page to recommend the new import path.

BREAKING CHANGE: `import '@adobe/spectrum-wc/badge'` no longer registers
`<swc-badge>` as a side effect. Use
`import '@adobe/spectrum-wc/components/badge/swc-badge.js'` instead.
