+++ 2nd-gen/packages/swc/components/asset/index.ts
@@ -9,14 +9,4 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { defineElement } from '@spectrum-web-components/core/element/index.js';
-
-import { Asset } from './Asset.js';
-
 export * from './Asset.js';
-declare global {
-  interface HTMLElementTagNameMap {
-    'swc-asset': Asset;
-  }
-}
-defineElement('swc-asset', Asset);
