+++ 2nd-gen/packages/swc/components/avatar/consumer-migration-guide.mdx
@@ -23,7 +23,7 @@ Update your import:
 import '@spectrum-web-components/avatar/sp-avatar.js';
 
 // After
-import '@adobe/spectrum-wc/avatar';
+import '@adobe/spectrum-wc/components/avatar/swc-avatar.js';
 ```
 
 > `@adobe/spectrum-wc` is a monolithic package. Importing via subpath (e.g. `@adobe/spectrum-wc/avatar`) loads only that component's bundle.
