+++ 2nd-gen/packages/swc/components/avatar/index.ts
@@ -9,14 +9,4 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { defineElement } from '@spectrum-web-components/core/element/index.js';
-
-import { Avatar } from './Avatar.js';
-
 export * from './Avatar.js';
-declare global {
-  interface HTMLElementTagNameMap {
-    'swc-avatar': Avatar;
-  }
-}
-defineElement('swc-avatar', Avatar);
