+++ 2nd-gen/packages/swc/components/avatar/swc-avatar.ts
@@ -9,4 +9,14 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-export * from './Divider.js';
+import { defineElement } from '@spectrum-web-components/core/element/index.js';
+
+import { Avatar } from './Avatar.js';
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'swc-avatar': Avatar;
+  }
+}
+
+defineElement('swc-avatar', Avatar);
