+++ 2nd-gen/packages/swc/components/avatar/test/avatar.test.ts
@@ -15,7 +15,7 @@ import type { Meta, StoryObj as Story } from '@storybook/web-components';
 
 import { Avatar } from '@adobe/spectrum-wc/avatar';
 
-import '@adobe/spectrum-wc/avatar';
+import '@adobe/spectrum-wc/components/avatar/swc-avatar.js';
 
 import { AVATAR_VALID_SIZES } from '../../../../core/components/avatar/Avatar.types.js';
 import {
