+++ 2nd-gen/packages/swc/components/badge/consumer-migration-guide.mdx
@@ -23,10 +23,10 @@ Update your import:
 import '@spectrum-web-components/badge/sp-badge.js';
 
 // After
-import '@adobe/spectrum-wc/badge';
+import '@adobe/spectrum-wc/components/badge/swc-badge.js';
 ```
 
-> `@adobe/spectrum-wc` is a monolithic package. Importing via subpath (e.g. `@adobe/spectrum-wc/badge`) loads only that component's bundle.
+> `@adobe/spectrum-wc` is a monolithic package. Importing via the per-component side-effect path (e.g. `@adobe/spectrum-wc/components/badge/swc-badge.js`) loads only that element.
 
 ## What changed
