+++ 2nd-gen/packages/swc/components/button/consumer-migration-guide.mdx
@@ -44,7 +44,7 @@ Replace `<sp-button>` with `<swc-button>`, update the import, and migrate rename
 // Before
 import '@spectrum-web-components/button/sp-button.js';
 // After
-import '@adobe/spectrum-wc/button';
+import '@adobe/spectrum-wc/components/button/swc-button.js';
 ```
 
 ### 2. Rename the tag and `treatment` attribute
