+++ 2nd-gen/packages/swc/components/button/index.ts
@@ -9,14 +9,4 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { defineElement } from '@spectrum-web-components/core/element/index.js';
-
-import { Button } from './Button.js';
-
 export * from './Button.js';
-declare global {
-  interface HTMLElementTagNameMap {
-    'swc-button': Button;
-  }
-}
-defineElement('swc-button', Button);
