+++ 2nd-gen/packages/swc/components/divider/consumer-migration-guide.mdx
@@ -23,7 +23,7 @@ Update your import:
 import '@spectrum-web-components/divider/sp-divider.js';
 
 // After
-import '@adobe/spectrum-wc/divider';
+import '@adobe/spectrum-wc/components/divider/swc-divider.js';
 ```
 
 > `@adobe/spectrum-wc` is a monolithic package. Importing via subpath (e.g. `@adobe/spectrum-wc/divider`) loads only that component's bundle.
