+++ 2nd-gen/packages/swc/components/icon/consumer-migration-guide.mdx
@@ -23,7 +23,7 @@ Update your import:
 import '@spectrum-web-components/icon/sp-icon.js';
 
 // After
-import '@adobe/spectrum-wc/icon';
+import '@adobe/spectrum-wc/components/icon/swc-icon.js';
 ```
 
 > `@adobe/spectrum-wc` is a monolithic package. Importing via subpath (e.g. `@adobe/spectrum-wc/icon`) loads only that component's bundle.
