+++ 2nd-gen/packages/swc/components/illustrated-message/consumer-migration-guide.mdx
@@ -41,7 +41,7 @@ Replace `<sp-illustrated-message>` with `<swc-illustrated-message>`, move headin
 import '@spectrum-web-components/illustrated-message/sp-illustrated-message.js';
 
 // After
-import '@adobe/spectrum-wc/illustrated-message';
+import '@adobe/spectrum-wc/components/illustrated-message/swc-illustrated-message.js';
 ```
 
 ### 2. Rename the tag
