+++ 2nd-gen/packages/swc/components/illustrated-message/index.ts
@@ -9,16 +9,4 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { defineElement } from '@spectrum-web-components/core/element/index.js';
-
-import { IllustratedMessage } from './IllustratedMessage.js';
-
 export * from './IllustratedMessage.js';
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'swc-illustrated-message': IllustratedMessage;
-  }
-}
-
-defineElement('swc-illustrated-message', IllustratedMessage);
