+++ 2nd-gen/packages/swc/components/progress-circle/consumer-migration-guide.mdx
@@ -23,7 +23,7 @@ Update your import:
 import '@spectrum-web-components/progress-circle/sp-progress-circle.js';
 
 // After
-import '@adobe/spectrum-wc/progress-circle';
+import '@adobe/spectrum-wc/components/progress-circle/swc-progress-circle.js';
 ```
 
 > `@adobe/spectrum-wc` is a monolithic package. Importing via subpath (e.g. `@adobe/spectrum-wc/progress-circle`) loads only that component's bundle.
