+++ 2nd-gen/packages/swc/components/progress-circle/index.ts
@@ -9,14 +9,4 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { defineElement } from '@spectrum-web-components/core/element/index.js';
-
-import { ProgressCircle } from './ProgressCircle.js';
-
 export * from './ProgressCircle.js';
-declare global {
-  interface HTMLElementTagNameMap {
-    'swc-progress-circle': ProgressCircle;
-  }
-}
-defineElement('swc-progress-circle', ProgressCircle);
