+++ 2nd-gen/packages/swc/components/status-light/consumer-migration-guide.mdx
@@ -23,7 +23,7 @@ Update your import:
 import '@spectrum-web-components/status-light/sp-status-light.js';
 
 // After
-import '@adobe/spectrum-wc/status-light';
+import '@adobe/spectrum-wc/components/status-light/swc-status-light.js';
 ```
 
 > `@adobe/spectrum-wc` is a monolithic package. Importing via subpath (e.g. `@adobe/spectrum-wc/status-light`) loads only that component's bundle.
