+++ 2nd-gen/packages/swc/components/status-light/index.ts
@@ -9,14 +9,4 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { defineElement } from '@spectrum-web-components/core/element/index.js';
-
-import { StatusLight } from './StatusLight.js';
-
 export * from './StatusLight.js';
-declare global {
-  interface HTMLElementTagNameMap {
-    'swc-status-light': StatusLight;
-  }
-}
-defineElement('swc-status-light', StatusLight);
