+++ 2nd-gen/packages/swc/components/tabs/index.ts
@@ -9,22 +9,6 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { defineElement } from '@spectrum-web-components/core/element/index.js';
-
-import { Tab } from './Tab.js';
-import { TabPanel } from './TabPanel.js';
-import { Tabs } from './Tabs.js';
-
 export * from './Tab.js';
 export * from './TabPanel.js';
 export * from './Tabs.js';
-declare global {
-  interface HTMLElementTagNameMap {
-    'swc-tab': Tab;
-    'swc-tab-panel': TabPanel;
-    'swc-tabs': Tabs;
-  }
-}
-defineElement('swc-tab', Tab);
-defineElement('swc-tab-panel', TabPanel);
-defineElement('swc-tabs', Tabs);
