+++ 2nd-gen/packages/swc/components/tabs/stories/tabs.stories.ts
@@ -13,7 +13,9 @@
 import { html, type TemplateResult } from 'lit';
 import type { Meta, StoryObj as Story } from '@storybook/web-components';
 
-import '@adobe/spectrum-wc/tabs';
+import '@adobe/spectrum-wc/components/tabs/swc-tabs.js';
+import '@adobe/spectrum-wc/components/tabs/swc-tab.js';
+import '@adobe/spectrum-wc/components/tabs/swc-tab-panel.js';
 
 import {
   KEYBOARD_ACTIVATIONS,
