+++ 2nd-gen/packages/swc/components/tabs/swc-tabs.ts
@@ -9,4 +9,14 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-export * from './Divider.js';
+import { defineElement } from '@spectrum-web-components/core/element/index.js';
+
+import { Tabs } from './Tabs.js';
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'swc-tabs': Tabs;
+  }
+}
+
+defineElement('swc-tabs', Tabs);
