+++ 2nd-gen/packages/swc/components/tabs/test/tabs.test.ts
@@ -15,7 +15,9 @@ import type { Meta, StoryObj as Story } from '@storybook/web-components';
 
 import { Tab, TabPanel, Tabs } from '@adobe/spectrum-wc/tabs';
 
-import '@adobe/spectrum-wc/tabs';
+import '@adobe/spectrum-wc/components/tabs/swc-tabs.js';
+import '@adobe/spectrum-wc/components/tabs/swc-tab.js';
+import '@adobe/spectrum-wc/components/tabs/swc-tab-panel.js';
 
 import { TABS_DIRECTIONS } from '../../../../core/components/tabs/Tabs.types.js';
 import {
