+++ 2nd-gen/packages/swc/patterns/conversational-ai/message-sources/MessageSources.ts
@@ -16,7 +16,7 @@ import { property } from 'lit/decorators.js';
 import { Chevron75Icon } from '@adobe/spectrum-wc/icon/elements/index.js';
 import { SpectrumElement } from '@spectrum-web-components/core/element/index.js';
 
-import '@adobe/spectrum-wc/icon';
+import '@adobe/spectrum-wc/components/icon/swc-icon.js';
 
 import { uniqueId } from '../../../utils/id.js';
