+++ 2nd-gen/packages/swc/patterns/conversational-ai/prompt-field/PromptField.ts
@@ -17,7 +17,7 @@ import { styleMap } from 'lit/directives/style-map.js';
 
 import { SpectrumElement } from '@spectrum-web-components/core/element/index.js';
 
-import '@adobe/spectrum-wc/icon';
+import '@adobe/spectrum-wc/components/icon/swc-icon.js';
 
 import { uniqueId } from '../../../utils/id.js';
 import { ChevronUpIcon, PlusIcon, StopIcon } from '../utils/icons/index.js';
