+++ 2nd-gen/packages/swc/patterns/conversational-ai/response-status/ResponseStatus.ts
@@ -17,8 +17,8 @@ import { ifDefined } from 'lit/directives/if-defined.js';
 import { Chevron75Icon } from '@adobe/spectrum-wc/icon/elements/index.js';
 import { SpectrumElement } from '@spectrum-web-components/core/element/index.js';
 
-import '@adobe/spectrum-wc/icon';
-import '@adobe/spectrum-wc/progress-circle';
+import '@adobe/spectrum-wc/components/icon/swc-icon.js';
+import '@adobe/spectrum-wc/components/progress-circle/swc-progress-circle.js';
 
 import { uniqueId } from '../../../utils/id.js';
 import { CheckCircleIcon } from '../utils/icons/index.js';
