+++ 2nd-gen/packages/swc/patterns/conversational-ai/upload-artifact/UploadArtifact.ts
@@ -15,7 +15,7 @@ import { property } from 'lit/decorators.js';
 
 import { SpectrumElement } from '@spectrum-web-components/core/element/index.js';
 
-import '@adobe/spectrum-wc/icon';
+import '@adobe/spectrum-wc/components/icon/swc-icon.js';
 
 import { CrossIcon } from '../utils/icons/index.js';
